jklm/requirements.txt
pyperclip
pyautogui
Pillow